Replace original EventFactory with the new one.
