+++ src/main/java/snw/kookbc/impl/KBCClient.java
@@ -35,7 +35,7 @@ import snw.kookbc.impl.console.Console;
 import snw.kookbc.impl.entity.builder.EntityBuilder;
 import snw.kookbc.impl.entity.builder.EntityUpdater;
 import snw.kookbc.impl.entity.builder.MessageBuilder;
-import snw.kookbc.impl.event.EventFactory2;
+import snw.kookbc.impl.event.EventFactory;
 import snw.kookbc.impl.network.Connector;
 import snw.kookbc.impl.network.HttpAPIRoute;
 import snw.kookbc.impl.network.NetworkClient;
@@ -64,7 +64,7 @@ public class KBCClient {
     private final EntityBuilder entityBuilder;
     private final MessageBuilder msgBuilder;
     private final EntityUpdater entityUpdater;
-    private final EventFactory2 eventFactory;
+    private final EventFactory eventFactory;
     private final ConfigurationSection config;
     private final File pluginsFolder;
     private final Session session = new Session(null);
@@ -86,7 +86,7 @@ public class KBCClient {
             @Nullable EntityBuilder entityBuilder,
             @Nullable MessageBuilder msgBuilder,
             @Nullable EntityUpdater entityUpdater,
-            @Nullable EventFactory2 eventFactory
+            @Nullable EventFactory eventFactory
     ) {
         if (pluginsFolder != null) {
             Validate.isTrue(pluginsFolder.isDirectory(), "The provided pluginsFolder object is not a directory.");
@@ -109,7 +109,7 @@ public class KBCClient {
         this.entityUpdater = Optional.ofNullable(entityUpdater).orElseGet(() -> new EntityUpdater(this));
         this.internalPlugin = new InternalPlugin(this);
         this.eventExecutor = Executors.newSingleThreadExecutor(r -> new Thread(r, "Event Executor"));
-        this.eventFactory = Optional.ofNullable(eventFactory).orElseGet(() -> new EventFactory2(this));
+        this.eventFactory = Optional.ofNullable(eventFactory).orElseGet(() -> new EventFactory(this));
     }
 
     // The result of this method can prevent the users to execute the console command,
@@ -350,7 +350,7 @@ public class KBCClient {
         return pluginMixinConfigManager;
     }
 
-    public EventFactory2 getEventFactory() {
+    public EventFactory getEventFactory() {
         return eventFactory;
     }
