+++ src/main/java/snw/kookbc/impl/network/ListenerImpl.java
@@ -32,7 +32,6 @@ import snw.jkook.message.component.TextComponent;
 import snw.kookbc.SharedConstants;
 import snw.kookbc.impl.KBCClient;
 import snw.kookbc.impl.command.CommandManagerImpl;
-import snw.kookbc.impl.event.EventFactory;
 import snw.kookbc.impl.network.exceptions.BadResponseException;
 import snw.kookbc.impl.network.webhook.WebHookClient;
 
@@ -137,7 +136,7 @@ public class ListenerImpl implements Listener {
     protected void event0(Frame frame) {
         Event event;
         try {
-            event = EventFactory.getEvent(client, frame);
+            event = client.getEventFactory().createEvent(frame.getData());
         } catch (Exception e) {
             client.getCore().getLogger().error("Unable to create event from payload.");
             client.getCore().getLogger().error("Event payload: {}", frame);
